+++ packages/core/src/components/_index.scss
@@ -29,7 +29,7 @@
 @import "non-ideal-state/non-ideal-state";
 @import "overflow-list/overflow-list";
 @import "overlay/overlay";
-@import "panel-stack2/panel-stack";
+@import "panel-stack/panel-stack";
 @import "popover/popover";
 @import "portal/portal";
 @import "progress-bar/progress-bar";
