+++ packages/core/src/components/panel-stack/_panel-stack.scss
