+++ packages/core/src/components/panel-stack/panel-stack.md
