+++ packages/core/src/components/panel-stack/panelStack.tsx
