+++ packages/core/src/components/panel-stack/panelTypes.ts
