+++ packages/core/src/components/panel-stack/panelView.tsx
